[WIP] clean up main
